refactor: :recycle: refactoring index.js
